Move state helpers out of barrel file.
