+++ ui/appui-layout-react/src/appui-layout-react/state/WidgetState.ts
@@ -65,72 +65,3 @@ export interface PopoutWidgetsState {
   readonly byId: { readonly [id: string]: PopoutWidgetState };
   readonly allIds: ReadonlyArray<PopoutWidgetState["id"]>;
 }
-
-/** @internal */
-export function addFloatingWidget(
-  state: NineZoneState,
-  id: FloatingWidgetState["id"],
-  tabs: WidgetState["tabs"],
-  floatingWidgetArgs?: Partial<FloatingWidgetState>,
-  widgetArgs?: Partial<WidgetState>
-): NineZoneState {
-  if (id in state.floatingWidgets.byId)
-    throw new UiError(category, "Floating widget already exists");
-
-  state = addWidgetState(state, id, tabs, widgetArgs);
-
-  let bounds = floatingWidgetArgs?.bounds;
-  if (!bounds) {
-    bounds = getNewFloatingWidgetBounds(state);
-  }
-
-  const tabId = tabs[0];
-  const tab = state.tabs[tabId];
-  const resizable = tab.isFloatingWidgetResizable;
-
-  const floatingWidget = createFloatingWidgetState(id, {
-    bounds,
-    resizable,
-    ...floatingWidgetArgs,
-  });
-  return produce(state, (stateDraft) => {
-    stateDraft.floatingWidgets.byId[id] = floatingWidget;
-    stateDraft.floatingWidgets.allIds.push(id);
-  });
-}
-
-/** @internal */
-export function addPopoutWidget(
-  state: NineZoneState,
-  id: PopoutWidgetState["id"],
-  tabs: WidgetState["tabs"],
-  popoutWidgetArgs?: Partial<PopoutWidgetState>,
-  widgetArgs?: Partial<WidgetState>
-): NineZoneState {
-  if (tabs.length !== 1)
-    throw new UiError(
-      category,
-      "Popout widget should contain one tab only",
-      undefined,
-      () => ({ tabs })
-    );
-
-  const popoutWidget = createPopoutWidgetState(id, popoutWidgetArgs);
-  state = addWidgetState(state, id, tabs, widgetArgs);
-  return produce(state, (stateDraft) => {
-    stateDraft.popoutWidgets.byId[id] = popoutWidget;
-    stateDraft.popoutWidgets.allIds.push(id);
-  });
-}
-
-/** @internal */
-export function floatingWidgetBringToFront(
-  state: NineZoneState,
-  floatingWidgetId: FloatingWidgetState["id"]
-): NineZoneState {
-  return produce(state, (draft) => {
-    const idIndex = draft.floatingWidgets.allIds.indexOf(floatingWidgetId);
-    const spliced = draft.floatingWidgets.allIds.splice(idIndex, 1);
-    draft.floatingWidgets.allIds.push(spliced[0]);
-  });
-}
