+++ ui/appui-layout-react/src/appui-layout-react/state/internal/PanelStateHelpers.ts
@@ -8,6 +8,7 @@
 
 import produce from "immer";
 import type { WritableDraft } from "immer/dist/types/types-external";
+import { UiError } from "@itwin/appui-abstract";
 import type { SizeProps } from "@itwin/core-react";
 import {
   type HorizontalPanelSide,
@@ -22,6 +23,9 @@ import type {
   PanelsState,
   VerticalPanelState,
 } from "../PanelState";
+import type { WidgetState } from "../WidgetState";
+import { category } from "./NineZoneStateHelpers";
+import { addWidgetState } from "./WidgetStateHelpers";
 
 function createPanelState(side: PanelSide) {
   return {
@@ -99,4 +103,38 @@ export function getPanelMaxSize(
   }
   const size = isHorizontalPanelSide(side) ? appSize.height : appSize.width;
   return (maxSize.percentage / 100) * size;
+}
+
+/** @internal */
+export function addPanelWidget(
+  state: NineZoneState,
+  side: PanelSide,
+  id: WidgetState["id"],
+  tabs: WidgetState["tabs"],
+  widgetArgs?: Partial<WidgetState>
+): NineZoneState {
+  return insertPanelWidget(state, side, id, tabs, Infinity, widgetArgs);
+}
+
+/** @internal */
+export function insertPanelWidget(
+  state: NineZoneState,
+  side: PanelSide,
+  id: WidgetState["id"],
+  tabs: WidgetState["tabs"],
+  sectionIndex: number,
+  widgetArgs?: Partial<WidgetState>
+): NineZoneState {
+  const panel = state.panels[side];
+  const maxWidgetCount = panel.maxWidgetCount;
+  if (panel.widgets.length >= maxWidgetCount)
+    throw new UiError(category, "Max widget count exceeded", undefined, () => ({
+      maxWidgetCount,
+    }));
+
+  state = addWidgetState(state, id, tabs, widgetArgs);
+  return produce(state, (draft) => {
+    const widgets = draft.panels[side].widgets;
+    widgets.splice(sectionIndex, 0, id);
+  });
 }
