+++ ui/appui-layout-react/src/test/state/TabLocation.test.ts
@@ -3,16 +3,15 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 import { should } from "chai";
+import { createNineZoneState, getTabLocation } from "../../appui-layout-react";
 import {
   addFloatingWidget,
-  addPanelWidget,
   addPopoutWidget,
-  createNineZoneState,
-  getTabLocation,
-  removeTabFromWidget,
-} from "../../appui-layout-react";
-import { addWidgetState } from "../../appui-layout-react/state/internal/WidgetStateHelpers";
+  addWidgetState,
+} from "../../appui-layout-react/state/internal/WidgetStateHelpers";
 import { addTabs } from "../Utils";
+import { addPanelWidget } from "../../appui-layout-react/state/internal/PanelStateHelpers";
+import { removeTabFromWidget } from "../../appui-layout-react/state/internal/TabStateHelpers";
 
 describe("getTabLocation", () => {
   it("should return floating tab location", () => {
