+++ ui/appui-layout-react/src/test/widget/MenuTab.test.tsx
@@ -12,8 +12,6 @@ import type {
   WidgetState,
 } from "../../appui-layout-react";
 import {
-  addPanelWidget,
-  addTab,
   createNineZoneState,
   ShowWidgetIconContext,
   TabIdContext,
@@ -23,6 +21,8 @@ import {
 } from "../../appui-layout-react";
 import { WidgetMenuTab } from "../../appui-layout-react/widget/MenuTab";
 import { TestNineZoneProvider } from "../Providers";
+import { addPanelWidget } from "../../appui-layout-react/state/internal/PanelStateHelpers";
+import { addTab } from "../../appui-layout-react/state/internal/TabStateHelpers";
 
 describe("MenuTab", () => {
   interface WrapperProps {
